requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
clang==5.0
cycler==0.11.0
cylinder-fitting==1.1.4
deap==1.3.1
deepdish==0.3.7
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==4.0.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.1.0
helpdev==0.7.1
idna==2.10
joblib==1.1.0
jupyter-packaging==0.10.4
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
laspy==2.0.2
Markdown==3.3.3
matplotlib==3.4.3
nibabel==3.2.1
nilearn==0.8.1
nltools==0.4.4
numpy==1.19.5
oauthlib==3.1.0
open3d==0.13.0
opencv-contrib-python==4.5.3.56
opencv-contrib-python-headless==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
Pillow==9.1.0
protobuf==3.20.0
ptyprocess==0.7.0
py-feat==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynv==0.2
pyoints==0.2.0a3
pyparsing==3.0.8
PyQtWebEngine==5.12.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.6.3
six==1.15.0
sklearn-genetic-opt==0.8.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.8.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.5.0
torchsummary==1.5.1
torchvision==0.6.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0